moved intersection story out of layer stories and wellbore in, made some tweaks to code
