+++ .storybook/src/layers/index.stories.ts
@@ -1,9 +1,8 @@
 export * from './grid-layer.stories';
 export * from './toggle-layer.stories';
 export * from './multiple-layers.stories';
-export * from '../wellborepath-layer.stories';
+export * from './wellborepath-layer.stories';
 export * from './callout-canvas.layer.stories';
-export * from './intersection.stories';
 
 export default {
   title: 'Layers',
