+++ src/layers/index.ts
@@ -4,3 +4,4 @@ export * from './GridLayer';
 export * from './SVGLayer';
 export * from './WellborePathLayer';
 export * from './CalloutCanvasLayer';
+export * from './ImageCanvasLayer';
